bug-fix( general information endpoints  ) and some general fixing

+++ app/Http/Controllers/API/GeneralController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\API;
 
 use App\Models\General;
 use App\Models\Language;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use App\Traits\APIResponseTrait;
 use App\Http\Controllers\Controller;
@@ -15,97 +16,90 @@ use App\Http\Requests\UpdateGeneralRequest;
 class GeneralController extends Controller
 {
     use APIResponseTrait;
+
     public function __construct()
     {
-        $this->middleware('auth:api',['except' => ['index','show']]);
+        $this->middleware('auth:api', ['except' => ['index', 'show']]);
     }
+
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request)
+    public function index(Request $request): JsonResponse
     {
         try {
-            $information = General::all();
+            $information = General::query();
             if ($request->header('language')) {
-                $language_header = $request->header('language');
-                $language = Language::where('name', '=', $language_header)->first();
+                if (!$language = Language::where('name', '=', $request->header('language'))->first())
+                    return $this->errorResponse('there are no language for this name');
 
-                $information = General::whereHas('langauges', function ($query) use ($language) {
-                    $query->where('language_id', '=', $language->id);
-                })->get();
+                $information->where('language_id', '=', $language->id)->get();
             }
-            $args['data'] = GeneralResource::collection($information);
-            return $this->successResponse($args , 200 );
+            return $this->successResponse(GeneralResource::collection($information->get()));
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreGeneralRequest $request)
+    public function store(StoreGeneralRequest $request): JsonResponse
     {
         try {
-            $validated = $request->validated();
-
-            $information =   General::create([
+            $information = General::create([
                 'name' => $request->name,
                 'value' => $request->value,
                 'language_id' => $request->language_id,
             ]);
             if ($request->hasFile('icon')) {
                 $icon = $request->file('icon');
-                $filename = time() . '.' .  $icon->getClientOriginalExtension();
+                $filename = time() . '.' . $icon->getClientOriginalExtension();
                 $icon->storeAs('public/images', $filename);
                 $information->icon = $filename;
                 $information->save();
             }
-            $args['message'] = 'general details  stored successfully ';
-            $args['data'] =new GeneralResource($information);
-            return $this->successResponse($args , 200 );
+            return $this->successOperationResponse('general details  stored successfully ');
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id, Request $request)
+    public function show(string $id, Request $request): JsonResponse
     {
         try {
-            $information = General::findOrFail($id);
+            $information = General::find($id);
+            if(!$information)
+                return $this->errorResponse('there are no information for this id ');
             if ($request->header('language')) {
-                $language_header = $request->header('language');
-                $language = Language::where('name', '=', $language_header)->first();
-                if ($language->id == $information->language_id) {
-                    $information = General::whereHas('langauges', function ($query) use ($language) {
-                        $query->where('language_id', '=', $language->id);
-                    })->first();
-                } else {
-                    return $this->FailResponse('go out');
-                }
+                if (!$language = Language::where('name', '=', $request->header('language'))->first())
+                    return $this->errorResponse('there are no language for this name');
+                if ($language->id != $information->language_id)
+                    return  $this->errorResponse('this information is for another language ');
             }
-            $args['data'] =new GeneralResource($information);
-            return $this->successResponse($args , 200 );
+            return $this->successResponse(new GeneralResource($information));
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateGeneralRequest $request, string $id)
+    public function update(UpdateGeneralRequest $request, string $id): JsonResponse
     {
         try {
-            $general = General::findOrFail($id);
+            $general = General::find($id);
+            if (!$general)
+                return $this->errorResponse('there are no info for this id ');
 
             if ($request->has('icon')) {
                 Storage::delete('public/images/' . $general->icon);
                 $icon = $request->file('icon');
-                $filename = time() . '.' .  $icon->getClientOriginalExtension();
+                $filename = time() . '.' . $icon->getClientOriginalExtension();
                 $icon->storeAs('public/images', $filename);
                 $general->icon = $filename;
             }
@@ -115,13 +109,12 @@ class GeneralController extends Controller
                 'language_id' => $request->language_id,
                 'icon' => $filename,
             ]);
-            $args['message'] = 'general details  updated successfully ';
-            $args['data'] =new GeneralResource($general);
-            return $this->successResponse($args , 200 );
+            return $this->successOperationResponse('general details  updated successfully ');
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
+
     /**
      * Remove the specified resource from storage.
      */
@@ -130,11 +123,9 @@ class GeneralController extends Controller
         try {
             Storage::delete('public/images/' . $general->icon);
             $general->delete();
-
-            $args['message'] = 'general details  deleted successfully ';
-            return $this->successResponse($args , 200);
+            return $this->successOperationResponse( 'general details  deleted successfully ');
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 }
